Calculate badges^ not authors

+++ src/pocketdb/repositories/ConsensusRepository.h
@@ -46,11 +46,11 @@ namespace PocketDb
     struct BadgeSharkConditions
     {
         int Height;
-        int LikersAll;
-        int LikersContent;
-        int LikersComment;
-        int LikersAnswer;
-        int RegistrationDepth;
+        int64_t LikersAll;
+        int64_t LikersContent;
+        int64_t LikersComment;
+        int64_t LikersAnswer;
+        int64_t RegistrationDepth;
     };
 
     struct BadgeSet
